requirements.txt
# Core Agent Framework
crewai==0.11.2
langchain>=0.1.0
langchain-community>=0.0.30

# LLM Providers
openai>=1.0.0
python-dotenv>=1.0.0

# Data & Utilities
pydantic>=2.0.0
requests>=2.31.0

# Logging & Monitoring
python-json-logger>=2.0.0

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
